Provide user option to override default logging
When calling zeus from a script that does its own logging, the fact that zeus relies on the root logger is somewhat vexing as it may result in duplicate logging. This commit exposes the option to pass an existing logger to the sampler so that the user can modify logging behaviour if they so desire. This also prevents zeus from clobbering handlers that have already been defined.